a2c/requirements.txt
torch>=2.3.0
numpy>=1.26.4
matplotlib>=3.9.0
torchvision>=0.18.0
scikit-image>=0.23.2
pillow>=10.3.0
tqdm>=4.66.4
gymnasium>=0.29.1
scikit-learn>=1.5.0
